driver_truck/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.3.1
python-decouple==3.8
requests==2.32.5
drf-spectacular==0.27.0
pyyaml==6.0.2